Cook sources location refactor
